Clean up the mmr history page so it isn't all living inside the player screen tsx
