+++ src/playerOverview/PlayerScreen.tsx
@@ -24,7 +24,6 @@ import {
     CategoryScale,
     LinearScale,
 } from 'chart.js';
-import { getMmrColor } from '../utils/mmrColorHelpers';
 import { SummonerCollage } from '../components/SummonerCollage';
 import {
     Flex,
@@ -51,6 +50,7 @@ import { Loading } from '../components/Loading';
 import { MmrHistoryItem } from '../types/domain/MmrHistoryItem';
 import { MmrCard } from '../components/MmrCard';
 import { FiChevronDown, FiChevronUp, FiMinus } from 'react-icons/fi';
+import { PlayerMmrSummary } from './PlayerMmrSummary';
 
 ChartJS.register(
     RadialLinearScale,
@@ -66,31 +66,6 @@ ChartJS.register(
 export async function loader(data: { params: any }) {
     return data.params.playerId;
 }
-
-const processMmrChanges = (data: MmrHistoryItem[]) => {
-    const playerMMR: { [key: string]: { gameId: number; mmr: number }[] } = {};
-
-    for (let i = 0; i < data.length; i++) {
-        const match = data[i];
-        for (const playerName of Object.keys(match.players)) {
-            if (playerMMR[playerName] === undefined) {
-                playerMMR[playerName] = [];
-            }
-
-            if (playerMMR[playerName].length > 0) {
-                let mostRecent =
-                    playerMMR[playerName][playerMMR[playerName].length - 1];
-                if (mostRecent.mmr === match.players[playerName]) continue;
-            }
-
-            playerMMR[playerName].push({
-                gameId: i + 1,
-                mmr: match.players[playerName],
-            });
-        }
-    }
-    return playerMMR;
-};
 
 /**
  * Given a player, create an array of champions that player has played with image url populated
@@ -135,39 +110,6 @@ export const PlayerScreen = React.memo(function PlayerScreen() {
 
     const matchHistoryResponse = ToxicDataService.useMatchHistory();
     const matchHistory = matchHistoryResponse.data ?? [];
-
-    const mmrPerMatchResponse = ToxicDataService.useMmrPerMatch();
-    const mmrPerMatch = mmrPerMatchResponse.data ?? [];
-    const mmrPerMatchMap = processMmrChanges(mmrPerMatch);
-    const playerMmrPerMatch = mmrPerMatchMap[playerId] ?? [];
-    const playerMmrPerMatchSliced = playerMmrPerMatch.slice(10).map((value) => {
-        return { x: value.gameId, y: value.mmr };
-    });
-    const last5Matches = playerMmrPerMatchSliced.slice(-5);
-    const mmrChangePercentage =
-        last5Matches.length > 0
-            ? Math.round(
-                  ((last5Matches[last5Matches.length - 1].y -
-                      last5Matches[0].y) /
-                      last5Matches[0].y) *
-                      10000
-              ) / 100
-            : 0;
-
-    const playerMmrPerMatchData = {
-        datasets: [
-            {
-                data: playerMmrPerMatchSliced,
-                fill: true,
-                backgroundColor: 'rgba(255, 99, 132, 0.2)',
-                borderColor: 'rgb(255, 99, 132)',
-                pointBackgroundColor: 'rgb(255, 99, 132)',
-                pointBorderColor: '#fff',
-                pointHoverBackgroundColor: '#fff',
-                pointHoverBorderColor: 'rgb(255, 99, 132)',
-            },
-        ],
-    };
 
     // only recompute the player classes when are looking at a new player
     const playerClasses = useMemo(
@@ -377,69 +319,7 @@ export const PlayerScreen = React.memo(function PlayerScreen() {
                         />
                     </TabPanel>
                     <TabPanel>
-                        <Flex direction={'column'}>
-                            <Flex
-                                direction={'row'}
-                                justifyContent={'center'}
-                                alignItems={'center'}
-                            >
-                                {playerMmrPerMatchSliced.length > 0 ? (
-                                    <>
-                                        <h1 style={{ fontSize: 60 }}>
-                                            {`${mmrChangePercentage}%`}
-                                        </h1>
-                                        {mmrChangePercentage == 0 ? (
-                                            <FiMinus
-                                                size={'60'}
-                                                color={'orange'}
-                                            />
-                                        ) : mmrChangePercentage > 0 ? (
-                                            <FiChevronUp
-                                                size={'60'}
-                                                color={'green'}
-                                            />
-                                        ) : (
-                                            <FiChevronDown
-                                                size={'60'}
-                                                color={'red'}
-                                            />
-                                        )}
-                                    </>
-                                ) : (
-                                    <h1 style={{ fontSize: 30 }}>
-                                        {`Not Placed! Player must complete 10 Games...`}
-                                    </h1>
-                                )}
-                            </Flex>
-                            {playerMmrPerMatchSliced.length > 0 ? (
-                                <Line
-                                    data={playerMmrPerMatchData}
-                                    style={{ maxHeight: 300 }}
-                                    options={{
-                                        scales: {
-                                            x: {
-                                                type: 'linear',
-                                                min: playerMmrPerMatchSliced[0]
-                                                    ?.x,
-                                                max: playerMmrPerMatchSliced[
-                                                    playerMmrPerMatchSliced.length -
-                                                        1
-                                                ]?.x,
-                                            },
-                                            y: {
-                                                suggestedMin: 1300,
-                                                suggestedMax: 1800,
-                                            },
-                                        },
-                                        plugins: {
-                                            legend: {
-                                                display: false,
-                                            },
-                                        },
-                                    }}
-                                />
-                            ) : null}
-                        </Flex>
+                        <PlayerMmrSummary playerId={player.name} />
                     </TabPanel>
                     <TabPanel>
                         <SortableTable
