--- src/utils/mmrColorHelpers.ts
@@ -1,3 +1,0 @@
-export function getMmrColor(mmr: number) {
-    return mmr > 0 ? `hsl(${120 * ((1800 - mmr) / 1800) * 4 * -1 + 120}, 100%, 67%)`: "transparent";
-}
